[PATCH] Be robust when "named argument" doesn't exist
**Summary**
The issue with the way positional args was defined is that the format conflicts with environment variables and nps was throwing an error when an environment variable was used, eg. consider a script that uses `$FOO`. When a token with the `$` prefix is not found in the args, just ignore it and continue.

**Test Plan**
- Updated tests