[PATCH] Change references to Markforged
**Summary**
There were many links to sezna's fork, and requests for funding that are no longer required to support the project. References to the old fork have been updated to the Markforged fork, and requests for funding were removed since Markforged is now supporting this project. A link to a paid for tutorial video was also removed since that did not feel appropriate.

A section was added to indicate that this package is being maintained by Markforged.

While here, I also cleaned up the many badges at the top to point to our CICD and remove things that seemed superfluous

**Test Plan**
- Links to the github repo can be output in certain cases and this is tested, so tests were updated